Autocomplete: rename logger to analytics logger (#5834)
- Renamed the autocomplete logger to an analytics logger for clarity.
The previous naming was confusing for those new to the autocomplete
section of the codebase, as it was unclear when to use `logDebug` and
when to use `CompletionsLogger.methodX`. We should establish a clearer
separation between utilities for debug output and events sent to
BigQuery and other third parties.
- I plan to create a separate debug output channel for autocomplete in a
follow-up, which would require changes around these "similar" function
calls.