Agent: re-enable bindings generation CI check
Fixes CODY-3809

Previously, we didn't automatically generate Kotlin bindings when making
changes to the agent protocol. This was problematic because it meant
that JetBrains, Visual Studio, and Eclipse teams need to manually write
the bindings to stay updated with the latest Cody.

Now, we re-enable the CI check. In addition, the codegen tool also
reports much more helpful errors now to increase the odds that a
developer can work around codegen issues in the future when making
changes to protocol types.

We have made a mistake in the past of leaking too many internal types
into the protocol. This is problematic because internal types tend to
lean heavily on advanced TypeScript typing techniques that are
complicated to translate to other languages. Instead, we should start
using more actively `Protocol*` named types, which mirror internal types
but don't rely on advanced type machinery. This is a bit extra work on
the agent side, but increases the odds we can keep the bindings
generation enabled in CI. Internal types are regulary updated with
features we don't need to care about in other IDE clients so this
technique should also lower the burden on VSC changes to constantly run
`pnpm generate-agent-kotlin-bindings`.