Chat: Display 'other' model category (#5902)
CLOSE :
https://linear.app/sourcegraph/issue/CODY-4059/opus-should-be-listed-under-othe

The issue: When processing the Opus model, which had a category of
'other', the getServerModelTags function was adding the
ModelTag.Balanced tag to it. This is why it was showing up under the
"Balanced" category in the UI.

The fix: This change adds a new 'other' model category to the ModelTag
enum and updates the related model service and configuration code to
handle this new category. This change allows the 'other' category to be
directly used without being converted to 'Balanced'.

The changes include:
- Adding 'other' as a new ModelTag enum value
- Updating the ModelCategory type to include 'other'
- Updating the getServerModelTags function to handle the 'other'
category
- Updating the mocked server models config to use the 'other' category
for the 'Claude 3 Opus' model

This change allows for more flexibility in categorizing models that
don't fit into the existing 'power', 'balanced', or 'speed' categories.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Added new unit tests for covering the model category.

Manual Test Plan:

Open the model drop down list to confirm Opus is listed under Other:


![image](https://github.com/user-attachments/assets/4b8d2473-78ec-48f5-8547-218a1038d93a)

Before:


![image](https://github.com/user-attachments/assets/a2b5ae65-beaa-4898-97ed-959a03cdfa02)


## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->