feat(uninstall): delete auth credentials on uninstall (CODY-1043) (#2434)
This is the JetBrains half of the
[CODY-1043](https://linear.app/sourcegraph/issue/CODY-1043/bug-account-credentials-are-cached-between-installs),
the VSC half of it is in [this
PR](https://github.com/sourcegraph/cody/pull/5819).

For VSCode I was able to write an E2E as there are electron & VSCode
test utilities for opening / closing instances of VSC and installing /
uninstalling plugins.

Does anyone know if we have similar utilities that we could use to write
a test for this behavior in JB?

The good news is that JB actually has proper APIs for this hook so the
complexity is much lower here.

## Test plan
I manually tested this by running `CODY_DIR='<my local cody>' ./gradlew
-PforceAgentBuild=true buildPlugin` and then opening the plugins menu
and installing a plugin from disk.

![image](https://github.com/user-attachments/assets/0b0c7482-9dce-4ce5-8f80-07b60857e1d9)

I logged in and then uninstalled from the same menu. After reinstalling,
it takes me back to the login screen and all previous endpoints are
cleared. When I log back into the same account I see my chat history is
removed.