Remove client-side query rewriting in ContextRetriever (#5904)
Previously, the `ContextRetriever` class was using a
`rewriteKeywordQuery` function to modify the original query before
sending it to the server for indexed context retrieval. This client-side
query rewriting is now redundant.

This PR removes the client-side query rewriting in the
`ContextRetriever` class. The changes include:

1. Removing the `rewriteKeywordQuery` function call from the
`ContextRetriever` class.
2. Passing the original query directly to the server for indexed context
retrieval.
3. Updating the `SymfRunner` class to use the original query for live
results.
4. Updating tests to reflect these changes and ensure the original query
is used throughout the retrieval process.
5. Updated and added tests in `ContextRetriever.test.ts` to verify the
correct usage of the original query.

The server-side now handles query rewriting for remote searches, making
the client-side rewriting unnecessary. This change simplifies the
client-side code, removes the needs of rewriting the search query
repeatedly.

symf is still using the `rewriteKeywordQuery` function in the getResults
method.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

- Updated existing unit tests to reflect the new behavior.
- Added new tests to ensure the original query is used in all relevant
methods.
- Manually tested to confirm that context retrieval still functions
correctly with these changes.

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->