adding autoedits support (#5845)
## Context
The PR enables a basic setup for `auto-edits`. The goal of the PR is for
us to check the model quality by manually triggering the `auto-edits`
and have a vibe check. The actual integration of next edits should be
done with the existing autocomplete infra. So, this PR is supposed to
live behind a feature flag just for the internal testing and model
iterations.

1. The PR takes 4 lines surrounding the cursor and than propose a diff
for those 4 lines.
2. The current implementation uses a fine-tuned gpt-4o-mini model. The
name of the model is:
`ft:gpt-4o-mini-2024-07-18:sourcegraph-production::AFXNjNiC`

## Steps to run the autoedits (in debug mode):
1. Set the following following setting in the vscode config
```
"cody.experimental.autoedits": {
    "provider": "openai",
    "model": "ft:gpt-4o-mini-2024-07-18:sourcegraph-production::AGgXey7l",
    "apiKey": "<openai_token>",
    "tokenLimit": {
      "prefixTokens": 2500,
      "suffixTokens": 2500,
      "maxPrefixLinesInArea": 12,
      "maxSuffixLinesInArea": 5,
      "codeToRewritePrefixLines": 2, 
      "codeToRewriteSuffixLines": 3,
      "contextSpecificTokenLimit": {
        "recent-view-port": 1000,
        "diagnostics": 1000,
        "recent-copy": 1000,
        "jaccard-similarity": 1000,
        "recent-edits": 1000
      }
    }
  },
```
2. Press `ctrl+shift+enter` to trigger the 
3. The debug console `Cody by Sourcegraph` will show the diff in the
console.
4. The suggestion will also be shown on the UI as a `ghost text` and we
can press `tab` to apply the changes and `escape` to reject the changes.


## Test plan
Updated CI checks and manual testing. Please see a demo below


https://github.com/user-attachments/assets/7c1cf50d-c1fa-48bc-a71e-364e184404ce

---------

Co-authored-by: Beatrix <beatrix@sourcegraph.com>