fix(webview): Skip webview copying for eclipse (#5924)
In the startup of the agent, we copy over the local webview assets into
the cody data directory. This overwrites the template files that eclipse
writes to that directory. The eclipse extension is packaged with the
webview assets it's compiled with and manages serving those resources
from it's own JAR. This change prevents any version mismatch issues that
can occur from the double copy.

Also includes a small fix for Java code generation.

## Test plan
Manually tested with Eclipse that it resolved issues loading CSS