added(git): Parse more exotic git URLs (CODY-3132) (#5512)
[CODY-3132](https://linear.app/sourcegraph/issue/CODY-3132/codys-convertgitcloneurltocodebasename-isnt-able-to-parse-some-git)

Adds support for git urls with subgroups or no owner, as well as port
numbers. However, I am unsure if these formatted URLs are used just for
printing or if they are accessed directly. I would love some insight on
how to test this.

Should only impact PLG users.

## Test plan
Added unit tests but I am curious to see if context generation is
affected and if more tests are needed.

## Changelog
Adds support for Git repositories with subgroups, ports, or no owner.