Move all feedback to community forum (#5923)
The feedback tickets on github is an extra entry point for us to filter
through and triage. I propose we update the template and funnel all
feedback to community for our team to bring forward instead. This way
our users will have our full attention and will not feel like we have
not actioned on any of the feedback.


## Test plan
N/A
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->