Add setContext notification for cody.serverEndpoint (#5918)
## Changes

I've added `cody.serverEndpoint` context variable, which on the clients
side is received through `window/didChangeContext` notification. We need
it on the jetbrains side in few places, e.g. when we are checking if
current account is a dotcom account.

I think it might be better idea to replace this with a dedicated
endpoint, but I'm not sure if `cody.activated` is currently used
anywhere?
@dominiccooney @sqs WDYT?



## Test plan

Full QA with https://github.com/sourcegraph/jetbrains/pull/2382/ is
needed.
It does not change anything on the Cody side.