Implement edit/insert prompts (#5958)
closes:
https://linear.app/sourcegraph/issue/SRCH-1172/prompts-can-output-into-3-places

This PR integrates with the prompt mode (chat | edit | insert) from the
SG API. Prompts with edit or insert mode are executed using the
`executeEdit`, the same way the commands are executed. The prompt is
used as the instruction text for `executeEdit`. First the chat
transcript is created, context is fetched and then the `executeEdit` is
called. The `FixupTask` from the execution is used to construct the
response in chat. The response for now shows the diff only.

In BG, the state for interactions mode is saved as intent. The dropdown
and icon indicator for to allow users to manually change the mode and to
see which mode is set, is only available behind the OneBox feature flag
and will not be released yet.

![CleanShot 2024-10-21 at 18 28
08@2x](https://github.com/user-attachments/assets/5cb3e6b1-b53a-41b8-824b-a74f35d1a4b9)

## Test plan
- create a edit/insert prompt
- execute it from Cody
- it should work as expected.

## Changelog

- Add ability to execute prompts to perform edits or insert code.