fix(chat): Hide insert and new file buttons if there is no `edit` capability (#6018)
Currently all clients default to showing the Insert and New file buttons
on code snippets in an assistant response, but only editors which
support the `edit` capability (i.e. not Eclipse or VS) should show these
buttons. This simply hides the action if the capability is not set to
`enabled`.

## Test plan
Tested this build in Eclipse and saw that only the Copy button was
retained, but on main, all three buttons were shown

## Changelog
Hide non-functional edit buttons in chat if client does not support the
functionality.