Apply chore improvements; update latest ide version (#2023)
~We need the GitHub releases and their links to look nice. A stable
release should not include `-nightly` in the link. This PR will
automatically push a new tag without the suffix so we will be able to
use it to generate the release notes.~

We cannot push from CI (without creating a token and setting it as a
secret). Let's stick to the present process (manually push stable tag
when needed).

Btw, I added a few minor improvements.

## Test plan
1. ~To test these changes we will need to do a stable release 😃~
1. N/A - these are only chore changes and improvements