feat(rel): add backport workflow (#6119)
This PR adds a backport github action based our backport tooling in
`sourcegraph/backport`
It uses different labels (`cherrypick vscode-v<version-number>`) which
is slightly different than the current label. Importantly, this allows
it to pick up the `base` branch to be used (in this case
`vscode-vX.YZ.x`) so it knows to backport to the branch.
Testing in this repo first, and if it works as anticipated, will port it
over to the JetBrains repo too.

## Test plan
Manually running this?
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

## Changelog
N/A
<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->

---------

Co-authored-by: Kalan <51868853+kalanchan@users.noreply.github.com>
(cherry picked from commit dd98f69d5f734690a500ee12ddd0c8dd1f229ee8)