autoedit: add speculative decoding (#6130)
## Context
Enable speculative decoding for the auto-edits deployment. The PR passes
the existing code we rewrite in the file as the speculative string.
Since only certain parts of the code that we rewrite change, most of the
tokens are present in the speculative string that we pass.

Backend PR: https://github.com/sourcegraph/sourcegraph/pull/1673 

## Test plan
Monitor the `Fireworks-Speculation-Matched-Tokens` header from
fireworks. This shows the number of tokens matched, which is linearly
related to prefix match between the code that we rewrite and response.

## Changelog
- autoedits: Enable speculative decoding.