Bench: make sure to respect CODY_RECORDING_MODE (#6167)
In context evals, we disable Polly recording so that responses to the
Sourcegraph backend are not cached. This is important, so we can test
local changes to the Sourcegraph backend and see updated eval results.

Before this PR, the command-bench code was accidentally ignoring
`CODY_RECORDING_MODE` and always starting a Polly recording.