DEEP CODY: remove user preferences on sync, update memory tool
This change updates the site preferences for the chat model to DEEP CODY on the first sync, instead of updating the user's model preferences. This ensures that the chat model is set to DEEP CODY for all users on the first sync, without modifying individual user preferences.

This PR also change initializes the CodyChatMemory singleton when the MemoryTool is constructed. This ensures that the chat memory is properly loaded and available for use in the chat context across chat session.

The CodyChatMemory class has also been updated to maintain a maximum of 8 most recent memory items in the static store, and to persist the memory items to local storage.

Remove github issue example from TOOLCLI to make sure it uses openCtx Tool when configured.