Fix crash on Edit Code  (#2699)
Closes
https://linear.app/sourcegraph/issue/QA-180/jetbrains-crash-occurred-while-clicking-on-edit-code-after-restarting.


## Test plan
See the linear issue.
You can try removing `AccountData` entry from `workspace.xml` before
opening the project.

## Changes

First I tried to simply setActiveAccount for the deprecated manager.
This is not the best solution as do not know some values for the account
(`name`, `displayName`, `id`). We could hardcode some values but we
would be persisting some artificial data for a simple task that is -
**all we need is to store the llm that was used for Edit Prompt most
recently**.

So instead of hacking & mocking the deprecated logic (HistoryService is
used basically only for history reasons, migrations, etc)... instead of
that, let's simply use a simply dedicated storage class.

Also, it could be an actual storage with a corresponding storage file
(.xml) but do we really need it? Project session persistence should the
enough for most users.