Update tracing for chat (#6230)
Make the following changes to tracing for chat:
    
- Removes the span tracking time-to-first-token. This didn't really make
sense to track as a separate span, since all it did was track the time
from the start of the chat handler to the first token emitted. This is
better tracked as an event on the root span.
- Add a span that tracks the duration of the LLM response (and records
an event for the first token)
- Add spans for other preprocessing actions, such as prompt construction
and context fetching.