Fixing Integrate OpenTelemetry tracing in Cody Webview PR 6100 (#6192)
#6100 was reverted in #6179 based[ on the thread
](https://sourcegraph.slack.com/archives/C05AGQYD528/p1732289326228309)
<img width="548" alt="image"
src="https://github.com/user-attachments/assets/a3d9e3f5-7b21-4779-85e4-321fb34e04e6">

So now @vovakulikov made a fix in this PR for it. The fix is in[ this
commit

](https://github.com/sourcegraph/cody/commit/b361e05f79d5fe0212bd3c2fd3dcf3e5866bf9a8)
NOTE: The commit is not present in this PR because of rebasing I had to
do some squash shenanigans but I made a separate branch where you can
see the commit which just shows the change from Vova.

## Test plan
- Go to web directory
- `pnpm dev` 
- go to http://localhost:5777 to test if the page loads properly and
everything else

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->