fix detecting the fireworks model (#6239)
## Context
With the introduction of new model configuration, the model identifiers
start with `provider::`. For fireworks, this would be
`fireworks::version::model_name`. This leads to missing a check when
using chat completions from sourcegraph instance to fireworks, which
always leads to appending the `assistant` message.
This leads to very bad quality for the fireworks auto-edit models which
shouldn't have this token.
The PR corrects the condition to detect the fireworks model.

## Test plan
Green CI