fix(agent): Client-managed secrets bug (#6238)
There is a bug in the current auth flow for client managed secrets. This
bug only affects Eclipse because it is the only extension using the
`client-managed` secrets capability. The issue is simply that the
`AgentClientManagedSecretStorage` did not fire the VSCode event when the
secret was updated, which caused the extension to use a cached stale
value for `AuthCredentials`.

## Test plan
Added unit tests.

## Changelog
Fixes a bug in the Eclipse plugin where the extension could become
unresponsive.